+++ main-app/service-worker.js
@@ -3,8 +3,8 @@
  * Caches all assets for offline use.
  */
 
-const APP_VERSION = 'v112';
-const CACHE_NAME = 'qr-cache-v112';
+const APP_VERSION = 'v113';
+const CACHE_NAME = 'qr-cache-v113';
 
 var ASSETS = [
   './',